feat(cacheable): cacheable support `as` attr
