+++ crates/rspack_cacheable_test/tests/macro/cacheable/mod.rs
@@ -1,0 +1,6 @@
+mod as_attr;
+mod basic;
+mod hashable;
+mod omit_bounds;
+mod with_attr;
+mod with_attr_and_generics;
